requirements.txt
flask
pandas==2.3.3
numpy==2.0.1
scikit-learn==1.7.2
xgboost==3.2.0
torch==2.5.1
torchvision==0.20.1
transformers==5.2.0
datasets==4.6.0
accelerate
evaluate
dill==0.4.0
Pillow
python-dotenv==1.2.1
pyarrow==23.0.1